+++ dart/glinty_flutter/lib/glinty_flutter.dart
@@ -9,3 +9,4 @@ export 'src/app.dart';
 export 'src/component.dart';
 export 'src/render.dart';
 export 'src/session.dart';
+export 'src/theme.dart';